取得件数一件の場合の処理を修正
